Source/Deployer/app_repo/test/requirements.txt
requests
flask
kafka-python
